+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * If QGIS 3.24, use the native function from QGIS server API to fetch the feature ID
 * Provide the Py-QGIS-Server version if possible in the JSON metadata
 * Use LRU Cache when reading the CFG file to avoid multiple access
+* Add Python, Qt and GDAL versions in the metadata API
 
 ## 1.0.0 - 2022-05-11